Update calcParent2Child.m
I don't think quantile on line 82 works very well. It cuts the sample used to get the new p2c by averaging to a really small size, which means that only those p2cs are used which are already very close to the current p2c. It's not hitting the modus of the orientation relationship distribution but instead a small sub-population that happens to be close to the starting p2c.

I also edited the convergence criterion to be a bit less stringent, so the modus actually gets through. There will always be a distribution of different orientation relationships and a needlessly tight requirement for good fit over the board basically makes the script spin its wheels endlessly.

Making these proposed edits resulted in a much, much improved OR for the example martensitic dataset, enabling the use of voting algorithms for prior austenite reconstruction.

I'm sure the situation can be quite different on a different steel and a different dataset. Perhaps calcParent2Child would require some more options?

-Tuomo